Bug 737667 - GCC build warning: sps_sampler.h: In member function ?void ProfileStack::push(const char*)?: warning: comparison between signed and unsigned integer expressions [-Wsign-compare] (also in addMarker) r=ehsan

+++ tools/profiler/sps_sampler.h
@@ -184,7 +184,7 @@ public:
     if (!aMarker) {
       return; //discard
     }
-    if (mMarkerPointer == mozilla::ArrayLength(mMarkers)) {
+    if (size_t(mMarkerPointer) == mozilla::ArrayLength(mMarkers)) {
       return; //array full, silently drop
     }
     mMarkers[mMarkerPointer] = aMarker;
@@ -213,7 +213,7 @@ public:
 
   void push(const char *aName)
   {
-    if (mStackPointer >= mozilla::ArrayLength(mStack)) {
+    if (size_t(mStackPointer) >= mozilla::ArrayLength(mStack)) {
       mDroppedStackEntries++;
       return;
     }
